+++ CHANGELOG.md
@@ -121,6 +121,10 @@
 
 #### Bug Fixes
 
+* Fix bug where SwiftLint ignores excluded files list in a nested configuration file.  
+  [Dylan Bruschi](https://github.com/Bruschidy54)
+  [#2447](https://github.com/realm/SwiftLint/issues/2447)
+
 * Fix false positives on `no_grouping_extension` rule when using `where`
   clause.  
   [Almaz Ibragimov](https://github.com/almazrafi)